refactored the copy link in trip summary into a component and made it so users can view trip in home page by using modal
