+++ src/components/TripDetails.jsx
@@ -1,42 +1,18 @@
 import React from "react";
 import { Card, Stack, Group, Button, Box, Title, Text, Modal, ActionIcon } from "@mantine/core";
 import { useDisclosure } from "@mantine/hooks";
-import DateSelector from "./DateSelector";
 import { IconCalendarWeek } from "@tabler/icons-react";
-import { notifications } from "@mantine/notifications";
+import CopyTripLink from "./CopyTripLink";
 
 
 
 const TripDetails = ({currTripId}) => {
     const [opened, { open, close }] = useDisclosure(false);
 
-    const handleLeaveTrip = () => {
+      const handleLeaveTrip = () => {
         console.log("Leaving trip (yes option was clicked)");
         close();
       };
-
-      const handleTripCopyLink = async () => {
-        try {
-          const inviteLink = `${window.location.origin}/tripsummary/${currTripId}`;
-      
-          await navigator.clipboard.writeText(inviteLink);
-          notifications.show({
-            title: "Link Copied!",
-            message: "Your invite link has been copied to the clipboard.",
-            color: "green",
-            autoClose: 3000,
-          });
-        } catch (err) {
-          console.error("Error copying invite link:", err);
-          notifications.show({
-            title: "Error",
-            message: "Could not copy invite link. Please try again.",
-            color: "red",
-            autoClose: 3000,
-          });
-        }
-      };
-      
 
   return (
     <Card shadow="sm" p="lg" radius="md" withBorder>
@@ -114,9 +90,7 @@ const TripDetails = ({currTripId}) => {
             </Text>
           </Box>
         </Stack>
-        <Button variant="light" fullWidth mt="md" onClick={handleTripCopyLink}>
-          Copy Link
-        </Button>
+        <CopyTripLink tripId={currTripId}/>
       </Stack>
       {/* Leave Trip Confirmation Modal */}
       <Modal opened={opened} onClose={close} title="Confirm Leave Trip" centered>
